proof of progress
I have attached proof of our assembly (the contigs) in the proofofprogress.txt and the I also attached a produced graph with the assembled contigs put into DeepARG pipeline which is used to predict the abundance of antimicrobial resistance genes. This is suppose to replicate figure 5 of the original pathofact paper, but in that paper, they did create figures with the psoriasis data. Therefore, this uploaded figure is not in the paper.